requirements.txt
sentence-transformers==2.2.2
langchain
ipykernel
flask
streamlit
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain_huggingface
